refactor: decompose AutomationChatView into focused components (FE-18)
Extract the 1,523-line monolith into 7 focused components + 1 composable,
each under 400 lines:

- ChatHeroHeader (119 lines) - page title, subtitle, action buttons
- LlmHealthStatusBar (188 lines) - provider health status banner
- ChatSessionSidebar (254 lines) - session list, create form, board selector
- ChatMessageList (361 lines) - message rendering with markdown support
- ChatParseHintCard (169 lines) - parse-hint instruction cards
- ChatToolCallDetails (135 lines) - tool call metadata expander
- ChatComposeBar (128 lines) - message input, proposal checkbox, send
- useAutomationChat composable (394 lines) - orchestration state/logic

Parent AutomationChatView reduced to 235 lines (template shell only).
All existing tests pass without modification.

Closes #859
